Better cross platform suppert

+++ src/input/softwareinterface.h
@@ -1,13 +1,12 @@
 #ifndef _ROBOT_INPUT_SOFTWAREINTERFACE_H_
 #define _ROBOT_INPUT_SOFTWAREINTERFACE_H_
 
-#include "value.h"
 
 namespace Robot::Input {
 
     class SoftwareInterface {
         public:
-            virtual void steer(double steer, double throttle, double aux_x=0.0, double aux_y=0.0) = 0;
+            virtual void steer(float steer, float throttle, float aux_x=0.0f, float aux_y=0.0f) = 0;
     };
 
 };
