+++ src/kinematic/controlscheme/frontwheel.cpp
@@ -28,7 +28,7 @@ ControlSchemeFrontWheel::~ControlSchemeFrontWheel()
 }
 
 
-void ControlSchemeFrontWheel::setMotors(double left, double right) 
+void ControlSchemeFrontWheel::setMotors(float left, float right) 
 {
     motorServo(FRONT_LEFT, Value::fromAngleRadians(WHEEL_STRAIGHT_ANGLE+left));
     motorServo(FRONT_RIGHT, -Value::fromAngleRadians(WHEEL_STRAIGHT_ANGLE+right));
