+++ src/kinematic/controlscheme/spinning.h
@@ -17,7 +17,7 @@ namespace Robot::Kinematic {
             virtual void init() override;
             virtual void cleanup() override;
 
-            virtual void steer(double steering, double throttle, double aux_x, double aux_y) override;
+            virtual void steer(float steering, float throttle, float aux_x, float aux_y) override;
 
         private:
     };
